use-cases/vertexai-pinecone/requirements.txt
pinecone>=5.4.0
google-auth>=2.38.0
requests>=2.32.0
opencv-python>=4.10.0
Pillow>=11.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-cloud-storage>=2.19.0
pypdf>=5.4.0
python-docx>=1.1.0
openai>=1.60.0  # optional, for audio transcription
